dev: extract normalizer options and context
